requirements-dev.txt
# ==============================================================================
# SM AI Backend - Development Dependencies (로컬 테스트용)
# ==============================================================================
# 로컬에서 API 서버 테스트 및 AI 분석 로직 검증
# ==============================================================================

# ------------------------------------------------------------------------------
# Core Framework
# ------------------------------------------------------------------------------
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# ------------------------------------------------------------------------------
# Database - PostgreSQL (읽기)
# ------------------------------------------------------------------------------
psycopg2-binary==2.9.9

# ------------------------------------------------------------------------------
# Database - MySQL (쓰기)
# ------------------------------------------------------------------------------
pymysql==1.1.1
cryptography==43.0.3

# ------------------------------------------------------------------------------
# AI - Gemini
# ------------------------------------------------------------------------------
google-generativeai==0.8.3

# ------------------------------------------------------------------------------
# Utility
# ------------------------------------------------------------------------------
requests==2.31.0
pandas==2.2.2
python-dateutil==2.9.0

# ------------------------------------------------------------------------------
# Testing
# ------------------------------------------------------------------------------
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==0.24.0      # FastAPI async 테스트
httpx==0.28.1               # FastAPI 테스트 클라이언트

# ------------------------------------------------------------------------------
# Development Tools
# ------------------------------------------------------------------------------
# black==24.10.0            # 코드 포매터
# pylint==3.3.3             # 린터
# mypy==1.14.0              # 타입 체커
